ovsdb: raft: Randomize leadership transfer.
Each cluster member typically always transfers leadership to the same
other member, which is the first in their list of servers.  This may
result in two servers in a 3-node cluster to transfer leadership to
each other and never to the third one.

Randomizing the selection to make the load more evenly distributed.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>