ovsdb: raft: Fix time intervals for multitasking while joining.
While joining, ovsdb-server may not wake up for a duration of a join
timer, which is 1 second and is by default 3x larger than a heartbeat
timer.  This is causing unnecessary warnings from the cooperative
multitasking module that thinks that we missed the heartbeat time by
a lot.

Use join timer instead while joining.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>