ovsdb: raft: Fix assertion when 1-node cluster looses leadership.
Some of the failure tests can make a single-node cluster to
loose leadership.  In this case the next raft_run() will
trigger election with a pre-vore enabled.  This is causing
an assertion when this server attempts to vote for itself.

Fix that by not using pre-voting if the is only one server.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>