ovsdb: raft: Avoid transferring leadership to unavailable servers.
Current implementation of the leadership transfer just shoots the
leadership in the general direction of the first stable server in the
configuration.  It doesn't check if the server was active recently or
even that the connection is established.  This may result in sending
leadership to a disconnected or otherwise unavailable server.

Such behavior should not cause log truncation or any other correctness
issues because the destination server would have all the append
requests queued up or the connection will be dropped by the leader.
In a worst case we will have a leader-less cluster until the next
election timer fires up.  Other servers will notice the absence of the
leader and will trigger a new leader election normally.
However, the potential wait for the election timer is not good as
real-world setups may have high values configured.

Fix that by trying to transfer to servers that we know have applied
the most changes, i.e., have the highest 'match_index'.  Such servers
replied to the most recent append requests, so they have highest
chances to be healthy.  Choosing the random starting point in the
list of such servers so we don't transfer to the same server every
single time.  This slightly improves load distribution, but, most
importantly, increases robustness of our test suite, making it cover
more cases.  Also checking that the message was actually sent without
immediate failure.

If we fail to transfer to any server with the highest index, try to
just transfer to any other server that is not behind majority and then
just any other server that is connected.  We did actually send them
all the updates (if the connection is open), they just didn't reply
yet for one reason or another.  It should be better than leaving the
cluster without a leader.

Note that there is always a chance that transfer will fail, since
we're not waiting for it to be acknowledged (and must not wait).
In this case, normal election will be triggered after the election
timer fires up.

Fixes: 1b1d2e6daa56 ("ovsdb: Introduce experimental support for clustered databases.")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>