netdev-dpdk: Unify Tx offload disabling for broken drivers.
Keeping track of all the drivers and different things that are not
working in them is not something OVS should do.  The fact that we even
check the driver names is already the thing that shouldn't exist.

It was somewhat manageable when it was one or two drivers, but it
seems like we're discovering more and more issues.

Instead of trying to keep up with all the drivers and their particular
issues, making a single list of drivers with known issues and just
disabling *all* Tx offloads for them.  Drivers can be added to the list
if new issues get uncovered and the list can be re-evaluated on DPDK
version updates.

'iavf' driver is added since it shares common issues with its PF
counterparts.  Bonding driver inherits incorrect flags from the
downstream drivers, so should always be on a list as log as there
is at least one other driver on that list.

Fixes: 084c8087292c ("userspace: Support VXLAN and GENEVE TSO.")
Fixes: 3337e6d91c5b ("userspace: Enable L4 checksum offloading by default.")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>