docs: Define Read The Docs configuration for Sphinx HTML parameters.
Read The Docs was always mangling the conf.py during the build to
inject custom domains configured in the project settings and some
other stuff.  But they will stop doing that soon [1].

Adding recommended changes to the config to get this info from the
environment.

Flake8 does't know that 'html_context' is a predefined dictionary,
so disabling the 'undefined name' warning.

[1] https://about.readthedocs.com/blog/2024/07/addons-by-default/

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>